Refactor project structure into folders for better organization
