+++ frontend-fiasco/src/components/CartItem.jsx
