+++ frontend-fiasco/src/components/Header.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import logo from "./assets/logo.jpg";
+import logo from "/src/assets/logo.jpg";
 
 function Header() {
     return (
