+++ frontend-fiasco/src/components/ProductItem.jsx
