+++ frontend-fiasco/src/layout/Layout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from "./Header";
+import Header from "../components/Header";
 import { useState } from "react";
 
 function Layout() {
